+++ Changelog.md
@@ -54,6 +54,7 @@
 * Ability to set handlers for command execution events (on_command_started/on_command_finished), 
   so that script setup/cleanup can be factorized, and script flow can be simplified, 
   avoiding try/finally blocks in order to restore defaults.
+* Setting data format and layout from command line (-dfmt, -dlay options).
 
   
 ### Changed